+++ .idea/compiler.xml
@@ -8,13 +8,13 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="paypal-demo" />
-        <module name="myappbackend" />
+        <module name="cupabakery-backend" />
       </profile>
     </annotationProcessing>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="myappbackend" options="-parameters" />
+      <module name="cupabakery-backend" options="-parameters" />
       <module name="paypal-demo" options="-parameters" />
     </option>
   </component>
