+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/myappbackend.iml" filepath="$PROJECT_DIR$/myappbackend.iml" />
+      <module fileurl="file://$PROJECT_DIR$/cupabakery-backend.iml" filepath="$PROJECT_DIR$/cupabakery-backend.iml" />
     </modules>
   </component>
 </project>
