+++ src/main/java/com/cupabakery/backend/MyAppBackend.java
@@ -1,0 +1,11 @@
+package com.cupabakery.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MyAppBackend{
+    public static void main(String[] args) {
+        SpringApplication.run(MyAppBackend.class, args);
+    }
+}
