+++ src/main/java/com/cupabakery/backend/config/PaypalProperties.java
@@ -1,0 +1,14 @@
+package com.cupabakery.backend.config;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Data
+@Configuration
+@ConfigurationProperties(prefix = "paypal")
+public class PaypalProperties {
+    private String clientId;
+    private String clientSecret;
+    private String mode;
+}
