+++ src/main/java/com/cupabakery/backend/controller/PaypalController.java
@@ -1,0 +1,35 @@
+package com.cupabakery.backend.controller;
+
+import com.cupabakery.backend.service.PaypalService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/paypal")
+@CrossOrigin(origins = "http://localhost:5173")
+@RequiredArgsConstructor
+public class PaypalController {
+
+    private final PaypalService paypalService;
+
+    @PostMapping("/create-paypal-order")
+    public Object createOrder() {
+        try {
+            String value = "50000";
+            String currency = "USD";
+            String reference = "DH" + System.currentTimeMillis();
+            return paypalService.createOrder(value, currency, reference);
+        } catch (Exception e) {
+            return e.getMessage();
+        }
+    }
+
+    @PostMapping("/capture-paypal-order")
+    public Object captureOrder(@RequestParam String orderId) {
+        try {
+            return paypalService.captureOrder(orderId);
+        } catch (Exception e) {
+            return e.getMessage();
+        }
+    }
+}
