+++ src/main/java/com/cupabakery/backend/dto/RoleDTO.java
@@ -1,0 +1,9 @@
+package com.cupabakery.backend.dto;
+
+import lombok.Data;
+
+@Data
+public class RoleDTO {
+    private int id;
+    private String name;
+}
