+++ src/main/java/com/cupabakery/backend/model/Amount.java
@@ -1,0 +1,13 @@
+package com.cupabakery.backend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Amount {
+    private String currency_code;
+    private String value;
+}
