+++ src/main/java/com/cupabakery/backend/model/AuthResponse.java
@@ -1,0 +1,13 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class AuthResponse {
+    private String scope;
+    private String access_token;
+    private String token_type;
+    private String app_id;
+    private int expires_in;
+    private String nonce;
+}
