+++ src/main/java/com/cupabakery/backend/model/CaptureOrderResponse.java
@@ -1,0 +1,15 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class CaptureOrderResponse {
+    private String id;
+    private String status;
+    private PaymentSource payment_source;
+    private List<PurchaseUnit> purchase_units;
+    private Payer payer;
+    private List<Link> links;
+}
