+++ src/main/java/com/cupabakery/backend/model/CreateOrderRequest.java
@@ -1,0 +1,15 @@
+package com.cupabakery.backend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateOrderRequest {
+    private String intent;
+    private List<PurchaseUnit> purchase_units;
+}
