+++ src/main/java/com/cupabakery/backend/model/CreateOrderResponse.java
@@ -1,0 +1,12 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class CreateOrderResponse {
+    private String id;
+    private String status;
+    private List<Link> links;
+}
