+++ src/main/java/com/cupabakery/backend/model/Link.java
@@ -1,0 +1,10 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class Link {
+    private String href;
+    private String rel;
+    private String method;
+}
