+++ src/main/java/com/cupabakery/backend/model/Name.java
@@ -1,0 +1,9 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class Name {
+    private String given_name;
+    private String surname;
+}
