+++ src/main/java/com/cupabakery/backend/model/Payer.java
@@ -1,0 +1,10 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class Payer {
+    private Name name;
+    private String email_address;
+    private String payer_id;
+}
