+++ src/main/java/com/cupabakery/backend/model/PaymentSource.java
@@ -1,0 +1,8 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class PaymentSource {
+    private Paypal paypal;
+}
