+++ src/main/java/com/cupabakery/backend/model/Paypal.java
@@ -1,0 +1,10 @@
+package com.cupabakery.backend.model;
+
+import lombok.Data;
+
+@Data
+public class Paypal {
+    private Name name;
+    private String email_address;
+    private String account_id;
+}
