+++ src/main/java/com/cupabakery/backend/model/PurchaseUnit.java
@@ -1,0 +1,13 @@
+package com.cupabakery.backend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PurchaseUnit {
+    private String reference_id;
+    private Amount amount;
+}
