+++ src/main/java/com/cupabakery/backend/model/Role.java
@@ -1,0 +1,19 @@
+package com.cupabakery.backend.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Data
+@Entity
+@Table(name = "roles")
+public class Role {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "role_id")
+    private int id;
+
+    @Column(name = "name", nullable = false, unique = true)
+    private String name;
+
+    // getters, setters, constructors
+}
