+++ src/main/java/com/cupabakery/backend/repository/RoleRepository.java
@@ -1,0 +1,9 @@
+package com.cupabakery.backend.repository;
+
+import com.cupabakery.backend.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByName(String name);
+}
+
