--- src/main/java/com/example/myappbackend/MyAppBackend.java
@@ -1,11 +1,0 @@
-package com.example.myappbackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MyAppBackend{
-    public static void main(String[] args) {
-        SpringApplication.run(MyAppBackend.class, args);
-    }
-}
