--- src/main/java/com/example/myappbackend/dto/RegisterDTO.java
@@ -1,34 +1,0 @@
-package com.example.myappbackend.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-
-@Data
-/*
- * Data transfer object for user register, validate the input data before --> Controller
- * 1. User submit -> formData -> RegisterDTO
- * 2. RegisterDTO validate the formData
- *    2.1 Valid data --> Controller
- *    2.2 Invalid data --> Throw error
-*/
-public class RegisterDTO {
-
-    @NotBlank(message = "Username không được để trống")
-    @Size(min = 3, max = 20, message = "Username phải có từ 3 đến 20 ký tự")
-    private String username;
-
-    @NotBlank(message = "Email không được để trống")
-    @Email(message = "Email không hợp lệ")
-    private String email;
-
-    @NotBlank(message = "Mật khẩu không được để trống")
-    @Size(min = 8, max = 40, message = "Mật khẩu phải có từ 8 đến 40 ký tự")
-    @Pattern(
-            regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!])(?=\\S+$).{8,}$",
-            message = "Mật khẩu phải chứa ít nhất 1 chữ hoa, 1 chữ thường, 1 số và 1 ký tự đặc biệt"
-    )
-    private String password;
-}
