--- src/main/java/com/example/myappbackend/exception/EmailExistsException.java
@@ -1,7 +1,0 @@
-package com.example.myappbackend.exception;
-
-public class EmailExistsException extends RuntimeException {
-    public EmailExistsException(String message) {
-        super(message);
-    }
-}
