--- src/main/java/com/example/myappbackend/exception/UsernameExistsException.java
@@ -1,7 +1,0 @@
-package com.example.myappbackend.exception;
-
-public class UsernameExistsException extends RuntimeException {
-    public UsernameExistsException(String message) {
-        super(message);
-    }
-}
