--- src/main/java/com/example/myappbackend/model/Amount.java
@@ -1,13 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Amount {
-    private String currency_code;
-    private String value;
-}
