--- src/main/java/com/example/myappbackend/model/AuthResponse.java
@@ -1,13 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-@Data
-public class AuthResponse {
-    private String scope;
-    private String access_token;
-    private String token_type;
-    private String app_id;
-    private int expires_in;
-    private String nonce;
-}
