--- src/main/java/com/example/myappbackend/model/CreateOrderRequest.java
@@ -1,15 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateOrderRequest {
-    private String intent;
-    private List<PurchaseUnit> purchase_units;
-}
