--- src/main/java/com/example/myappbackend/model/CreateOrderResponse.java
@@ -1,12 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class CreateOrderResponse {
-    private String id;
-    private String status;
-    private List<Link> links;
-}
