--- src/main/java/com/example/myappbackend/model/Link.java
@@ -1,10 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-@Data
-public class Link {
-    private String href;
-    private String rel;
-    private String method;
-}
