--- src/main/java/com/example/myappbackend/model/Name.java
@@ -1,9 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-@Data
-public class Name {
-    private String given_name;
-    private String surname;
-}
