--- src/main/java/com/example/myappbackend/model/PaymentSource.java
@@ -1,8 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-@Data
-public class PaymentSource {
-    private Paypal paypal;
-}
