--- src/main/java/com/example/myappbackend/model/Paypal.java
@@ -1,10 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.Data;
-
-@Data
-public class Paypal {
-    private Name name;
-    private String email_address;
-    private String account_id;
-}
