--- src/main/java/com/example/myappbackend/model/PurchaseUnit.java
@@ -1,13 +1,0 @@
-package com.example.myappbackend.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PurchaseUnit {
-    private String reference_id;
-    private Amount amount;
-}
