+++ src/main/resources/application.properties
@@ -18,3 +18,8 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+
+spring.jmx.enabled=false
+
+# Setting dev environment
+spring.profiles.active=dev
